Making improvements
